+++ lib/src/presentation/screens/splash/splash_screen.dart
@@ -9,7 +9,7 @@ import '../../theme/colors.dart';
 import '../../viewmodels/checkin_view_model.dart';
 import '../../viewmodels/privacy_lock_view_model.dart';
 
-const String _kAppVersion = '1.1.14';
+const String _kAppVersion = '1.1.16';
 
 class SplashScreen extends ConsumerStatefulWidget {
   const SplashScreen({super.key});